moved spotlight pass code into spotlight.cpp
