+++ src/lights/dirlight.h
@@ -34,40 +34,6 @@
 #define BBL 6
 #define FBL 7
 
-#if 0
-struct spotlight_t {
-  int id = -1;
-  transform_t transform;
-  float radius = 1.f;
-  static const float NEAR_PLANE;
-  static const float FAR_PLANE;
-  static const float SHADOW_MAP_WIDTH;
-  static const float SHADOW_MAP_HEIGHT;
-  vec3 dir;
-  vec3 color;
-
-#if SHOW_LIGHTS
-  static int LIGHT_MESH_ID;
-#endif
-  static shader_t light_shader;
-  framebuffer_t light_pass_fb;
-
-  mat4 view;
-  mat4 proj;
-};
-
-void init_light_data();
-int create_spotlight(vec3 pos);
-void setup_spotlight_for_rendering(int light_id);
-void remove_spotlight_from_rendering();
-spotlight_t get_spotlight(int light_id);
-int get_num_spotlights();
-tex_id_t get_spotlight_fb_depth_tex(int light_id);
-mat4 get_spotlight_proj_mat(int light_id);
-mat4 get_spotlight_view_mat(int light_id);
-vec3 get_spotlight_pos(int light_id);
-#endif
-
 struct dir_light_shadow_map_vert_t {
   vec2 pos;
   vec2 tex;
